revert src/cluster.cpp, changes made have no functional change.

+++ src/cluster.cpp
@@ -8,7 +8,7 @@ https://github.com/robertmartin8/RandomWalks/blob/master/kmeans.cpp
 #include <cmath>
 #include <stdexcept>
 #include <iostream>
-#include <iomanip>
+
 #include "point.hpp"
 #include "cluster.hpp"
 
@@ -21,14 +21,11 @@ https://github.com/robertmartin8/RandomWalks/blob/master/kmeans.cpp
  */
 bool calcMinimumDistances(std::vector<Point>* points, std::vector<Point>* centroids) {
     bool changed = false;
-    //TODO: MPI broadcast centroids?
-    //TODO: MPI scatter points?
-    //TODO: MPI gather points?
     for (std::vector<Point>::iterator centroidIterator = centroids->begin(); centroidIterator != centroids->end(); centroidIterator++) {
         int clusterId = centroidIterator - centroids->begin();
         for (std::vector<Point>::iterator pointIterator = points->begin(); pointIterator != points->end(); pointIterator++) {
             Point point = *pointIterator;
-            double distance = centroidIterator->distance(point);
+            float distance = centroidIterator->distance(point);
             if (distance < point.minDistance) {
                 //update the point's centroid (what cluster it belongs to)
                 point.minDistance = distance;
@@ -51,9 +48,7 @@ bool calcMinimumDistances(std::vector<Point>* points, std::vector<Point>* centro
 void moveCentroids(std::vector<Point>* points, std::vector<Point>* centroids, int k) {
     //Create vectors to keep track of data needed to compute means
     std::vector<int> numberOfPointsInEachCluster(k, 0);
-    std::vector<std::vector<double>> sums(points->at(0).coordinates.size());
-    //TODO: MPI gather? reduce? -- deprioritize
-    //TODO: swap index order for sums -- cluster first then dimension?
+    std::vector<std::vector<float>> sums(points->at(0).coordinates.size());
     for (int j = 0; j < k; j++) {
         for (size_t d = 0; d < sums.size(); d++) {
             sums[d].push_back(0.0);
@@ -62,8 +57,6 @@ void moveCentroids(std::vector<Point>* points, std::vector<Point>* centroids, in
 
     //Compute means
     //Compute sum of coordinates per cluster for each dimension
-    // TODO: MPI scatter points and centroids? don't gather sums?
-    // TODO: note -- do not scatter/reduce whatever on dimensions (d), that will not be efficient for the dataset.
     for (std::vector<Point>::iterator pointIterator = points->begin(); pointIterator != points->end(); pointIterator++) {
         int clusterId = pointIterator->cluster;
         numberOfPointsInEachCluster[clusterId] += 1;
@@ -71,9 +64,7 @@ void moveCentroids(std::vector<Point>* points, std::vector<Point>* centroids, in
             sums[d][clusterId] += pointIterator->coordinates[d];
         }
     }
-
     //Move centroids to the mean coordinate of the points in its cluster
-    //TODO: MPI scatter centroids, gather editted centroids?
     for (std::vector<Point>::iterator centroidIterator = centroids->begin(); centroidIterator != centroids->end(); centroidIterator++) {
         int clusterId = centroidIterator - centroids->begin();
         for (size_t d = 0; d < sums.size(); d++) {
@@ -118,16 +109,6 @@ void kMeansCluster(std::vector<Point>* points, int maxEpochs, int k){
     for (int epoch = 0; epoch < maxEpochs; epoch++) {
         // compute the distance from each centroid to each point
         // update the point's cluster as necessary.
-//        std::cout <<"SERIAL EPOCH " << epoch << " Centroids:\n";
-//        for (int i = 0; i < k; i++){
-//            std::cout << "  Centroid " << i << ": (";
-//            for (size_t d = 0; d < centroids.at(i).coordinates.size(); d++){
-//                std::cout << std::fixed << std::setprecision(6) << centroids.at(i).coordinates[d];
-//                if (d< centroids.at(i).coordinates.size() -1) std::cout << ", ";
-//            }
-//            std::cout << ")\n";
-//        }
-//        std::cout << std::endl;
         bool changed = calcMinimumDistances(points, &centroids);
         moveCentroids(points, &centroids, k);
         if(changed == false){
